Fix comparison
-> Eliminate duplication
